requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
arxiv==1.4.8
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
deprecation==2.1.0
exceptiongroup==1.1.3
feedparser==6.0.10
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gspread==5.10.0
gspread-dataframe==3.3.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
pdfminer.six==20221105
pikepdf==8.4.1
Pillow==10.0.1
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.1.1
pypdf==3.16.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
yarl==1.9.2
